bpo-43950: Make compiler output more source offsets.
This PR is part of PEP 657 and augments the compiler to emit ending
line numbers as well as starting and ending columns from the AST
into compiled code objects. This allows bytecodes to be correlated
to the exact source code ranges that generated them.

This information is made available through the following public APIs:

* The `co_positions` method on code objects.
* The C APIs `PyCode_Addr2EndLine`, `PyCode_Addr2Offset` and
  `PyCode_Addr2EndOffset`

Co-authored-by: Batuhan Taskaya <isidentical@gmail.com>
Co-authored-by: Pablo Galindo <Pablogsal@gmail.com>